requirements.txt
django==3.2
django-recurrence==1.11.1
huey==2.4.3
redis==4.3.4
freezegun==1.2.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
dj-rest-auth==2.2.5
django-allauth==0.51.0
responses==0.21.0 # for tests
django-filter==22.1
django-cors-headers==3.13.0